+++ justlib/src/ft_strslen.c
@@ -13,9 +13,9 @@
 /* ************************************************************************** */
 #include "justlib.h"
 
-size_t	ft_strslen(const char **s)
+int	ft_strslen(const char **s)
 {
-	size_t		i;
+	int		i;
 
 	i = 0;
 	while (s[i])
